bump mempool proto version to v2 and move new proto messages to v2
