+++ api/cometbft/mempool/v2/message.go
@@ -1,4 +1,4 @@
-package v1
+package v2
 
 import (
 	"fmt"
@@ -13,12 +13,29 @@ func (m *Txs) Wrap() proto.Message {
 	return mm
 }
 
+func (m *HaveTx) Wrap() proto.Message {
+	mm := &Message{}
+	mm.Sum = &Message_HaveTx{HaveTx: m}
+	return mm
+}
+
+func (m *ResetRoute) Wrap() proto.Message {
+	mm := &Message{}
+	mm.Sum = &Message_ResetRoute{ResetRoute: m}
+	return mm
+}
+
 // Unwrap implements the p2p Wrapper interface and unwraps a wrapped mempool
 // message.
 func (m *Message) Unwrap() (proto.Message, error) {
 	switch msg := m.Sum.(type) {
 	case *Message_Txs:
 		return m.GetTxs(), nil
+	case *Message_HaveTx:
+		return m.GetHaveTx(), nil
+	case *Message_ResetRoute:
+		return m.GetResetRoute(), nil
+
 	default:
 		return nil, fmt.Errorf("unknown message: %T", msg)
 	}
