+++ proto/cometbft/mempool/v1/types.proto
@@ -8,23 +8,10 @@ message Txs {
   repeated bytes txs = 1;
 }
 
-// HaveTx is sent by the DOG protocol to signal a peer that the sender already
-// has a transaction.
-message HaveTx {
-  bytes tx_key = 1;
-}
-
-// ResetRoute is sent by the DOG protocol to signal a peer to reset a (random)
-// route to the sender.
-message ResetRoute {
-}
-
 // Message is an abstract mempool message.
 message Message {
   // Sum of all possible messages.
   oneof sum {
     Txs txs = 1;
-    HaveTx have_tx = 2;
-    ResetRoute reset_route = 3;
   }
 }
