+++ proto/cometbft/mempool/v2/types.proto
@@ -1,0 +1,30 @@
+syntax = "proto3";
+package cometbft.mempool.v2;
+
+option go_package = "github.com/cometbft/cometbft/api/cometbft/mempool/v2";
+
+// Txs contains a list of transaction from the mempool.
+message Txs {
+  repeated bytes txs = 1;
+}
+
+// HaveTx is sent by the DOG protocol to signal a peer that the sender already
+// has a transaction.
+message HaveTx {
+  bytes tx_key = 1;
+}
+
+// ResetRoute is sent by the DOG protocol to signal a peer to reset a (random)
+// route to the sender.
+message ResetRoute {
+}
+
+// Message is an abstract mempool message.
+message Message {
+  // Sum of all possible messages.
+  oneof sum {
+    Txs txs = 1;
+    HaveTx have_tx = 2;
+    ResetRoute reset_route = 3;
+  }
+}
